Captain's log, stardate 96208.85 (8/10/18 at 3:17 PM),
After a long, hard struggle, the Enterprise and its crew have managed to survive a close encounter with a strange monster. For over a week and a half, a bug was caught in the engine that prevented the moving parts from hitting one another properly. After every attempt to eliminate this creature, it grew stronger, and our methods grew more and more convoluted. However, thanks to our Sourcetree Warp Drive, we were able to warp back far enough to allow our engineers to eliminate the beast before it became a problem.